reduce particle count on mobile
